doc pass and default cleanup
Contains readme cleanup and removal of the CI project name for a project variable in the examples.

In addition I'm allowing project services to be disabled on destroy is causing trouble with CI because we aren't properly sanitizing each run by using separate ephemeral projects. Since that isn't an option we're going to just leave the project services on.